+++ backend-test/Controllers/ControllerTests.hs
@@ -8,11 +8,11 @@ Module that contains the test suites for all the controllers.
 module Controllers.ControllerTests
 (  controllerTests  ) where
 
-import Test.HUnit (Test (..))
+import Test.Tasty (TestTree, testGroup)
 
 import Controllers.CourseControllerTests (courseControllerTestSuite)
 import Controllers.GraphControllerTests (graphControllerTestSuite)
 
 -- Single test encompassing all controller test suites
-controllerTests :: Test
-controllerTests = TestList [courseControllerTestSuite, graphControllerTestSuite]
+controllerTests :: TestTree
+controllerTests = testGroup "Controller" [courseControllerTestSuite, graphControllerTestSuite]