+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### 🔧 Internal changes
 
 - Refactored navigation bar into a React component (for the graph, grid and generate pages only - the about page navigation bar is still rendered using Blaze)
+- Refactored backend tests to use `tasty` and `tasty-hunit` instead of `HUnit`
 
 ## [0.7.1] - 2025-06-16